Atualização de Versão para 1.2.0 Perfil vinculado a permissão. Cadastro de perfil. Historico de compra. Correções de layout

+++ Nextt_Gestao_Compra/Nextt_Gestao_Compra.Dominio/Entidades/Atributos.cs
@@ -7,7 +7,7 @@ namespace Nextt_Gestao_Compra.Dominio.Entidades
         public Int16 IDTipoAtributo { get; set; }
         public Int16 IDTipoAtributoKey { get; set; }
         public Byte CasaDecimal { get; set; }
-        public Byte Ordem { get; set; }
+        public Int16 Ordem { get; set; }
         public Byte Tipo { get; set; }
         public bool Ativo { get; set; }
         public bool MultiplaSelecao { get; set; }
